Changed blast-bwa benchmark to use medium data instead of small.
Small took a bit over 4 hours when run sequentially on one node.  This is too
small a test for a 600 core cluster running all in parallel.  Medium is 200
times the size of small.  The medium test has 200000 sequences, at 4 lines per
sequence.  Simply cut it down using head, so that there are 4 times as many
lines in the file as sequences desired.